Fixed Bug #{3027}
Done!

Included with the third party software are datatypes that isolate its API calls from the underlying machines characteristics. The Tcl binding needs to use those, and it does... except for one case.

Since JNEM is being built as a true 64 bit application, this will be hard to test in I&T. For now, we'll just have to wait until Athena starts using geo-referenced maps to test this.

+++ src/libGeoTiff/geoTiffRead.cpp
@@ -366,7 +366,7 @@ geoTiffReader_getGeoFieldCmd (ClientData cd, Tcl_Interp *interp, int objc,
                                Tcl_Obj* CONST objv[])
 {
     double *d_list = NULL;
-    int     d_list_count;
+    uint16  d_list_count;
     ttag_t  field;
 
     GeoTiffInfo* gti = (GeoTiffInfo*)cd;
